AddItems was cleaned and reorganized C is new file for constants that can be written there and are easy to use Shopiing product is was changed in need of double numbers instead of Int numbers Content view was added a feature to sort list or unsort list EditItems will be view for editing items Hold for More view is prototype simples way of adding items Shopping list  is was changed in need of double numbers instead of Int numbers ShoppingListApp was changed just i need of testing HoldForMoreView

+++ ShoppingList/Model/ShoppingList.swift
@@ -11,7 +11,7 @@ struct ShoppingList : Identifiable, Equatable, Codable, Hashable {
     var id = UUID()
     var item = String()
     var category: Categories
-    var number = Int()
+    var number = Double()
     var value = String()
     var store: StoreName
     var isBought = Bool()
